requirements.txt
requests>=2.28.0
websocket-client>=1.6.0
sounddevice>=0.4.6
numpy>=1.24.0
edge-tts>=6.1.0
